refactor(security): extract LazyJwtToken class to security-core module : mutualize it instead of duplicating this class
