+++ gravitee-gateway-security/gravitee-gateway-security-core/pom.xml
@@ -37,5 +37,11 @@
             <artifactId>gravitee-gateway-policy</artifactId>
             <version>${project.version}</version>
         </dependency>
+
+        <dependency>
+            <groupId>com.nimbusds</groupId>
+            <artifactId>nimbus-jose-jwt</artifactId>
+            <version>${nimbus-jwt.version}</version>
+        </dependency>
     </dependencies>
 </project>
